// types/admission.ts
export type AdmissionData = {
  studentName: string;
  fatherName?: string;
  dateOfBirth?: string;
  caste?: string;
  profession?: string;
  previousEducation?: string;
  previousSchool?: string;
  class?: string;
  homePhone?: string;
  personalPhone?: string;
  dateOfEntry?: string;
  address?: string;
  photoUrl?: string;
};